app/drf/requirements.txt
asgiref==3.3.4
boto3==1.17.48
botocore==1.20.48
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-restframework==0.0.1
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4